分离主进程window
