+++ src/renderer/index.ts
@@ -2,11 +2,13 @@ import {createApp} from "vue";
 import App from "./App.vue";
 import {argsSymbol, createArgs} from "./store";
 import router from "./router";
-import {Init} from "./utils/ipc";
+import {windowLoad, messageBack} from "./utils/ipc";
 
-Init().then((Args) => {
+(async () => {
+    messageBack();
+    const Args = await windowLoad();
     createApp(App as any)
         .use(router)
         .provide(argsSymbol, createArgs(Args))
         .mount("#app");
-});
+})()
